With new dataset types coming online there are several (geochron and charcoal) that produce responses with no count data (we consider charcoal 'lab data').
In the case of `geochronology` records there is a specific API for these records that actually returns a response.  The `download` for a geochronology returns the dataset and  empty slots.  For this reason `get_download` now returns a message and NULL for `geochronologies`.  The NULL is then stripped from the final response.

`charcoal` data is a bit more complicated.  For now it is returned as a `download` with empty count data, but the download has no chronological information associated with it.  This is as a result of the API and data structure in neotoma itself and can't otherwise be solved in a straightforward manner.

To accomodate the empty count data and empty chron data there are now two `if` statements, one at 207 and one at 306 that help us resolve the problem.

`print.download` also had to be modified to accomodate data without any associated chronological information.