Moved the `gpid` test down now that the `gpid` term accepts both numeric and character values.
Added a single line into `get_dataset` to address a strange API issue where calls with `gpid` only are returning a different variable than any other `dataset` call.  This might be only temporary, but Neotoma uses `CollType` and not `CollUnitType` in the actual SQL structure, so we're going to keep it in.