Added code of conduct.  Rebuilt package.  Version change with changes in the packages used.  Completely removed dependencies on `RJSONIO` and `RCurl`.
`bind` - Clearly defining if statements to ensure that the outside case (not a defined class) gets caught properly for a warning.

`get_chroncontrol` - added the option to add the chron control to the associated `download` object as part of the list (~line 120).

`get_table` removed deprecated function call in the example.

`print.dataset` - cleaned up printing method to account for NA access dates if a record is read in from Tilia (e.g.).

Some new tests.