Found some bugs in `read.tilia` when going through a set of example (and incomplete) files from the neotoma holding tank.
Some of this is minor, and this will be itterative.  One key problem was the failure of the method to accurately capture some of the dataset information.  I've cleaned up some of the xml2 code.

The other problem was records that did not have age models/chronologies yet.  This has also been fixed.